move slugify to its own file
